movies/app/src/test/java/com/app/movies/ui/viewModel/MainActivityViewModelTest.kt
package com.app.movies.ui.viewM

import com.app.movies.ui.viewModel.MainActivityViewModel
import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.Observer
import com.app.movies.RxImmediateSchedulerRule
import com.app.movies.model.MovieData
import com.app.movies.model.Movies
import com.app.movies.network.APIService
import io.reactivex.Observable
import org.junit.*
import org.junit.runner.RunWith
import org.mockito.Mock
import org.mockito.Mockito
import org.mockito.MockitoAnnotations
import org.mockito.junit.MockitoJUnitRunner

@RunWith(MockitoJUnitRunner.Silent::class)

class MainActivityViewModelTest {

    // A JUnit Test Rule that swaps the background executor used by
    // the Architecture Components with a different one which executes each task synchronously.
    // You can use this rule for your host side tests that use Architecture Components.
    @Rule @JvmField var rule = InstantTaskExecutorRule()

    // Test rule for making the RxJava to run synchronously in unit test
    companion object {
        @ClassRule @JvmField val schedulers = RxImmediateSchedulerRule()
    }

    @Mock lateinit var apiService: APIService

    @Mock private lateinit var movieObseraval: Observable<Movies>

    @Mock private lateinit var movieObserver: Observer<Movies>

    lateinit var mainActivityViewModel: MainActivityViewModel


    @Before
    fun setUp() {
        MockitoAnnotations.initMocks(this)
        mainActivityViewModel = MainActivityViewModel(apiService)
    }


    @Test
    fun `given success state, when movie called, then update live data for success status`() {
        Mockito.`when`(apiService.fetchMovies()).thenReturn(movieObseraval)

        // observe on the MutableLiveData with an observer
        mainActivityViewModel.movies.observeForever(movieObserver)
        mainActivityViewModel.getMovieList()
        val movieObseraval = mainActivityViewModel.fetchMovieListObserver()

        Assert.assertEquals(movieObseraval.value, mainActivityViewModel.movies.value)
        mainActivityViewModel.loading.value?.let { Assert.assertTrue(it) }
    }

    @Test
    fun `given error state, when movie called, then show error for error status`() {
        movieObseraval = Observable.error(Throwable())

        Mockito.`when`(apiService.fetchMovies()).thenReturn(movieObseraval)
        mainActivityViewModel.getMovieList()

        mainActivityViewModel.postLoadError.value?.let { Assert.assertTrue(it.isNotBlank()) }
    }

    @Test
    fun `given search text, when fetch filter movie called, if match then return filtered movie list`() {
        val searchText = Observable.just("Search")

        val movieData = MovieData("search", 0, "200", "search", "200")
        val filterMovieList: ArrayList<MovieData> = ArrayList()
        filterMovieList.add(movieData)
        filterMovieList.add(movieData)
        filterMovieList.add(movieData)

        Mockito.`when` (mainActivityViewModel.setQueryListener(searchText)).thenReturn(Unit)



        val filterMovieListn: ArrayList<MovieData> = ArrayList()
        for (it in filterMovieList) {
            if (it.genre.contains(
                    searchText.toString(), true
                ) && (it.title.contains(searchText.toString(), false))
            ) {
                filterMovieListn.add(it)
            }
        }
        val expectedList: MutableLiveData<Movies> = MutableLiveData()
        expectedList.postValue(Movies(filterMovieListn))
        mainActivityViewModel.setQueryListener(searchText)
        val actualList = mainActivityViewModel.fetchFilterList()


        //Assert.assertEquals(expectedList,actualList)
        Assert.assertTrue(expectedList!=null)
        //mainActivityViewModel.postLoadError.value?.let { Assert.assertTrue(it.isNotBlank()) }
    }

}
